+++ order-service/src/main/java/com/nikea/productservice/service/logic/pricestrategy/impl/ChairPriceStrategy.java
@@ -14,7 +14,7 @@ public class ChairPriceStrategy extends AbstractPriceStrategy {
     }
 
     @Override
-    public int calculatePrice(double price) {
+    public double calculatePrice(double price) {
         return random((int) Math.round(price));
     }
 }
