+++ order-service/src/main/java/com/nikea/productservice/service/logic/pricestrategy/impl/TablePriceStrategy.java
@@ -16,7 +16,7 @@ public class TablePriceStrategy extends AbstractPriceStrategy {
 
 
     @Override
-    public int calculatePrice(double price) {
+    public double calculatePrice(double price) {
         return random((int) (FurnitureType.TABLE.ordinal() + price));
     }
 }
